@@ -18,48 +18,40 @@
   "publishConfig": {
     "access": "public"
   },
-  "main": "index.js",
+  "type": "module",
   "files": [
-    "jest.js",
-    "react.js",
-    "playwright.js",
-    "vitest.js"
+    "lib"
   ],
+  "exports": {
+    ".": {
+      "types": "./lib/index.d.ts",
+      "import": "./lib/index.js"
+    }
+  },
   "scripts": {
     "lint": "eslint .",
     "fmt": "prettier --write ."
   },
   "peerDependencies": {
-    "@typescript-eslint/eslint-plugin": ">= 5",
-    "@typescript-eslint/parser": ">= 5",
-    "eslint": ">= 8",
-    "eslint-plugin-jest": ">= 25",
-    "eslint-plugin-react": ">= 7",
-    "eslint-plugin-react-hooks": ">= 4",
-    "eslint-plugin-sort": ">= 2.10.0",
-    "eslint-plugin-vitest": ">= 0.2.8"
-  },
-  "peerDependenciesMeta": {
-    "eslint-plugin-react": {
-      "optional": true
-    },
-    "eslint-plugin-react-hooks": {
-      "optional": true
-    },
-    "eslint-plugin-jest": {
-      "optional": true
-    },
-    "eslint-plugin-vitest": {
-      "optional": true
-    }
+    "eslint": ">= 9"
   },
   "dependencies": {
-    "eslint-config-prettier": "^8.3.0"
+    "@eslint/js": "^9.13.0",
+    "@typescript-eslint/eslint-plugin": "^8.11.0",
+    "@typescript-eslint/parser": "^8.11.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-jest": "^28.8.3",
+    "eslint-plugin-playwright": "^1.8.1",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "eslint-plugin-sort": "^3.1.0",
+    "eslint-plugin-vitest": "^0.5.4",
+    "globals": "^15.11.0",
+    "typescript-eslint": "^8.11.0"
   },
   "devDependencies": {
     "@mskelton/semantic-release-config": "^1.0.1",
-    "eslint": "^8.49.0",
-    "eslint-plugin-sort": "^2.10.0",
+    "eslint": "^9.13.0",
     "prettier": "^3.0.3",
     "semantic-release": "^23.0.2"
   }
