@@ -1,6 +1,12 @@
-module.exports = {
-  extends: "plugin:playwright/playwright-test",
+// @ts-check
+
+import playwright from "eslint-plugin-playwright"
+
+/** @type {import("eslint").Linter.Config} */
+export default {
+  ...playwright.configs["flat/recommended"],
   rules: {
+    ...playwright.configs["flat/recommended"].rules,
     "playwright/missing-playwright-await": [
       "error",
       { customMatchers: ["toPassAxe"] },
