@@ -1,15 +1,22 @@
-module.exports = {
-  env: {
-    browser: true,
-  },
-  extends: "plugin:react/recommended",
-  parserOptions: {
-    ecmaFeatures: {
-      jsx: true,
+import react from "eslint-plugin-react"
+import reactHooks from "eslint-plugin-react-hooks"
+import globals from "globals"
+
+/** @type {import("eslint").Linter.Config} */
+export default {
+  ...react.configs.flat.recommended,
+  languageOptions: {
+    ...react.configs.flat.recommended.languageOptions,
+    globals: {
+      ...globals.browser,
     },
   },
-  plugins: ["react-hooks"],
+  plugins: {
+    ...react.configs.flat.recommended.plugins,
+    "react-hooks": reactHooks,
+  },
   rules: {
+    ...react.configs.flat.recommended.rules,
     "react/button-has-type": "error",
     "react/destructuring-assignment": "error",
     "react/function-component-definition": [
